Feature/Card Created Card Component
