+++ BE/Components/Card/Card.cpp
@@ -1,0 +1,3 @@
+#include "Card.h"
+#include "pch.h"
+#include <string>
