+++ BE/Components/Card/CardEnums.h
@@ -1,0 +1,10 @@
+#pragma once
+
+enum class CardTypes
+{
+	Empty = 0,
+	Clubs = 1,
+	Diamonds = 2,
+	Hearts = 3,
+	Spades = 4,
+};
