+++ FiveCards.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 16.0.33529.622
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "FiveCards", "FiveCards.vcxproj", "{86D1FF25-D268-415C-B375-5159E843E59C}"
 EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Card", "BE\Components\Card\Card.vcxproj", "{DC87C778-FC91-4887-8C60-EC5255F17972}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|x64 = Debug|x64
@@ -21,6 +23,14 @@ Global
 		{86D1FF25-D268-415C-B375-5159E843E59C}.Release|x64.Build.0 = Release|x64
 		{86D1FF25-D268-415C-B375-5159E843E59C}.Release|x86.ActiveCfg = Release|Win32
 		{86D1FF25-D268-415C-B375-5159E843E59C}.Release|x86.Build.0 = Release|Win32
+		{DC87C778-FC91-4887-8C60-EC5255F17972}.Debug|x64.ActiveCfg = Debug|x64
+		{DC87C778-FC91-4887-8C60-EC5255F17972}.Debug|x64.Build.0 = Debug|x64
+		{DC87C778-FC91-4887-8C60-EC5255F17972}.Debug|x86.ActiveCfg = Debug|Win32
+		{DC87C778-FC91-4887-8C60-EC5255F17972}.Debug|x86.Build.0 = Debug|Win32
+		{DC87C778-FC91-4887-8C60-EC5255F17972}.Release|x64.ActiveCfg = Release|x64
+		{DC87C778-FC91-4887-8C60-EC5255F17972}.Release|x64.Build.0 = Release|x64
+		{DC87C778-FC91-4887-8C60-EC5255F17972}.Release|x86.ActiveCfg = Release|Win32
+		{DC87C778-FC91-4887-8C60-EC5255F17972}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
